feat: LectureSchedule 엔티티 추가
