+++ src/main/java/hhplus/lectures/domain/LectureScheduleRepository.java
@@ -1,0 +1,4 @@
+package hhplus.lectures.domain;
+
+public interface LectureScheduleRepository {
+}
