+++ src/main/java/hhplus/lectures/infra/persistence/LectureScheduleJpaRepository.java
@@ -1,0 +1,7 @@
+package hhplus.lectures.infra.persistence;
+
+import hhplus.lectures.domain.LectureSchedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LectureScheduleJpaRepository extends JpaRepository<LectureSchedule, Long> {
+}
