+++ src/main/java/hhplus/lectures/infra/persistence/LectureScheduleRepositoryImpl.java
@@ -1,0 +1,6 @@
+package hhplus.lectures.infra.persistence;
+
+import hhplus.lectures.domain.LectureScheduleRepository;
+
+public class LectureScheduleRepositoryImpl implements LectureScheduleRepository {
+}
